Move readers and writers to one file
